requirements.txt
langchain
openai
faiss-cpu
tiktoken
pypdf
streamlit
flask
